backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
flask-cors
bcrypt
psycopg2-binary
gunicorn
prophet
scikit-learn
pandas
apscheduler
joblib
